Split observability status modules
